Removes unneeded casts

+++ Antumbra/Utility/Mixer.cs
@@ -10,9 +10,9 @@ namespace Antumbra.Glow.Utility {
         public static Color16Bit MixColorPercIn(Color16Bit newColor, Color16Bit prevColor, double newWeight) {
             double prevWeight = 1.00 - newWeight;
             Color16Bit result;
-            result.red = Convert.ToUInt16((int)(prevColor.red * prevWeight) + (int)(newColor.red * newWeight));
-            result.green = Convert.ToUInt16((int)(prevColor.green * prevWeight) + (int)(newColor.green * newWeight));
-            result.blue = Convert.ToUInt16((int)(prevColor.blue * prevWeight) + (int)(newColor.blue * newWeight));
+            result.red = Convert.ToUInt16(prevColor.red * prevWeight + newColor.red * newWeight);
+            result.green = Convert.ToUInt16(prevColor.green * prevWeight + newColor.green * newWeight);
+            result.blue = Convert.ToUInt16(prevColor.blue * prevWeight + newColor.blue * newWeight);
             return result;
         }
     }
